requirements.txt
# Core dependencies for LLM Analysis Benchmark
openai>=1.12.0
google-generativeai>=0.3.2
anthropic>=0.18.0
requests>=2.31.0

# Configuration and data handling
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Analysis and benchmarking
numpy>=1.24.0
pandas>=2.0.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Performance monitoring
psutil>=5.9.0